+++ src/main.rs
@@ -89,6 +89,7 @@ async fn main() -> io::Result<()> {
             .service(semaphore_service::block_until_acquired)
             .service(semaphore_service::remove_expired)
             .service(semaphore_service::put_peer)
+            .service(semaphore_service::is_acquired)
             .service(semaphore_service::freeze)
             .default_service(
                 // 404 for GET requests