+++ src/semaphore_service.rs
@@ -135,6 +135,13 @@ async fn remainder(
     state.remainder(&query.semaphore).map(Json)
 }
 
+/// Returns wether all the locks of the peer have been acquired. This route will not block, but
+/// return immediatly.
+#[get("/peers/{id}/is_acquired")]
+async fn is_acquired(path: Path<u64>, state: Data<State>) -> Result<Json<bool>, ThrottleError> {
+    state.is_acquired(*path).map(Json)
+}
+
 #[delete("/peers/{id}")]
 async fn release(path: Path<u64>, state: Data<State>) -> HttpResponse {
     if state.release(*path) {